Revert "multiapps"
This reverts commit eb850d70d92b25bdfbae078c07a0cbf6564b9f5c.
