+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/handlers/v2/ConfigurationParser.java
@@ -13,30 +13,25 @@ import com.sap.cloud.lm.sl.mta.schema.SchemaValidator;
 public class ConfigurationParser {
 
     protected final SchemaValidator platformValidator;
-    
+
     public ConfigurationParser() {
         this(new SchemaValidator(Schemas.PLATFORM));
     }
 
-    protected ConfigurationParser(SchemaValidator platformsValidator) {
-        this.platformValidator = platformsValidator;
+    protected ConfigurationParser(SchemaValidator platformValidator) {
+        this.platformValidator = platformValidator;
     }
 
     public Platform parsePlatformJson2(String json) throws ParsingException {
-        return (Platform) parsePlatformJson(json);
+        return parsePlatform(JsonUtil.convertJsonToMap(json));
     }
 
-    public Platform parsePlatformJson(String json) throws ParsingException {
+    public Platform parsePlatformJson2(InputStream json) throws ParsingException {
         return parsePlatform(JsonUtil.convertJsonToMap(json));
     }
     
-    private Platform parsePlatform(Map<String, Object> source) {
-        platformValidator.validate(source);
-        return getPlatformParser(source).parse();
-    }
-    
-    public Platform parsePlatformJson2(InputStream json) throws ParsingException {
-        return (Platform) parsePlatformJson(json);
+    public Platform parsePlatformJson(String json) throws ParsingException {
+        return parsePlatform(JsonUtil.convertJsonToMap(json));
     }
 
     public Platform parsePlatformJson(InputStream json) throws ParsingException {
@@ -46,6 +41,11 @@ public class ConfigurationParser {
         } catch (IOException e) {
             throw new IllegalStateException(e.getMessage(), e);
         }
+    }
+
+    private Platform parsePlatform(Map<String, Object> source) {
+        platformValidator.validate(source);
+        return getPlatformParser(source).parse();
     }
 
     protected PlatformParser getPlatformParser(Map<String, Object> source) {
