+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/handlers/v3/DescriptorHandler.java
@@ -2,6 +2,7 @@ package com.sap.cloud.lm.sl.mta.handlers.v3;
 
 import static com.sap.cloud.lm.sl.common.util.CommonUtil.cast;
 
+import com.sap.cloud.lm.sl.mta.handlers.ModulesSorter;
 import com.sap.cloud.lm.sl.mta.model.v3.DeploymentDescriptor;
 import com.sap.cloud.lm.sl.mta.model.v3.ExtensionDescriptor;
 import com.sap.cloud.lm.sl.mta.model.v3.ExtensionRequiredDependency;
@@ -66,9 +67,8 @@ public class DescriptorHandler extends com.sap.cloud.lm.sl.mta.handlers.v2.Descr
     }
 
     @Override
-    protected com.sap.cloud.lm.sl.mta.handlers.v2.ModulesSorter getModuleSorter(
-        com.sap.cloud.lm.sl.mta.model.v2.DeploymentDescriptor descriptor, String parallelDeploymentProperty, String dependencyTypeProperty,
-        String hardDependencyType) {
+    protected ModulesSorter getModuleSorter(com.sap.cloud.lm.sl.mta.model.v2.DeploymentDescriptor descriptor,
+        String parallelDeploymentProperty, String dependencyTypeProperty, String hardDependencyType) {
         DeploymentDescriptor descriptorV3 = cast(descriptor);
         return new com.sap.cloud.lm.sl.mta.handlers.v3.ModulesSorter(descriptorV3, this, dependencyTypeProperty, hardDependencyType,
             parallelDeploymentProperty);
