+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/model/v2/ExtensionProvidedDependency.java
@@ -15,6 +15,7 @@ import com.sap.cloud.lm.sl.mta.model.Visitor;
 
 public class ExtensionProvidedDependency implements VisitableElement, NamedElement, PropertiesContainer {
 
+    //TODO: implement ParametersContainer ?
     private String name;
     private Map<String, Object> properties;
