+++ com.sap.cloud.lm.sl.mta/src/test/java/com/sap/cloud/lm/sl/mta/builders/ExtensionDescriptorChainBuilderTest.java
@@ -12,8 +12,8 @@ import org.junit.jupiter.api.Test;
 
 import com.sap.cloud.lm.sl.common.ContentException;
 import com.sap.cloud.lm.sl.mta.message.Messages;
-import com.sap.cloud.lm.sl.mta.model.v2.DeploymentDescriptor;
-import com.sap.cloud.lm.sl.mta.model.v2.ExtensionDescriptor;
+import com.sap.cloud.lm.sl.mta.model.v1.DeploymentDescriptor;
+import com.sap.cloud.lm.sl.mta.model.v1.ExtensionDescriptor;
 
 public class ExtensionDescriptorChainBuilderTest {
