+++ com.sap.cloud.lm.sl.mta/src/test/java/com/sap/cloud/lm/sl/mta/builders/v2/ParametersChainBuilderTest.java
@@ -10,7 +10,7 @@ import com.sap.cloud.lm.sl.mta.handlers.v2.DescriptorParser;
 import com.sap.cloud.lm.sl.mta.model.v2.DeploymentDescriptor;
 import com.sap.cloud.lm.sl.mta.model.v2.Platform;
 
-public class ParametersChainBuilderTest extends PropertiesChainBuilderTest {
+public class ParametersChainBuilderTest extends com.sap.cloud.lm.sl.mta.builders.v1.PropertiesChainBuilderTest {
 
     @Parameters
     public static Iterable<Object[]> getParameters() {
@@ -50,7 +50,8 @@ public class ParametersChainBuilderTest extends PropertiesChainBuilderTest {
         });
     }
 
-    public ParametersChainBuilderTest(String deploymentDescriptorLocation, String platformLocation, Expectation[] expectations) {
+    public ParametersChainBuilderTest(String deploymentDescriptorLocation, String platformLocation,
+        Expectation[] expectations) {
         super(deploymentDescriptorLocation, platformLocation, expectations);
     }
 
@@ -65,7 +66,8 @@ public class ParametersChainBuilderTest extends PropertiesChainBuilderTest {
     }
 
     @Override
-    protected PropertiesChainBuilder createPropertiesChainBuilder(DeploymentDescriptor deploymentDescriptor, Platform platform) {
+    protected PropertiesChainBuilder createPropertiesChainBuilder(
+        com.sap.cloud.lm.sl.mta.model.v1.DeploymentDescriptor deploymentDescriptor, com.sap.cloud.lm.sl.mta.model.v1.Platform platform) {
         return new ParametersChainBuilder((DeploymentDescriptor) deploymentDescriptor, (Platform) platform);
     }
