+++ com.sap.cloud.lm.sl.mta/src/test/java/com/sap/cloud/lm/sl/mta/handlers/v2/ConfigurationParserTest.java
@@ -1,29 +1,13 @@
 package com.sap.cloud.lm.sl.mta.handlers.v2;
 
-import java.io.InputStream;
 import java.util.Arrays;
 
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 
-import com.sap.cloud.lm.sl.common.util.Callable;
-import com.sap.cloud.lm.sl.common.util.TestUtil;
 import com.sap.cloud.lm.sl.common.util.TestUtil.Expectation;
-import com.sap.cloud.lm.sl.mta.model.v2.Platform;
 
-@RunWith(Parameterized.class)
-public class ConfigurationParserTest {
+public class ConfigurationParserTest extends com.sap.cloud.lm.sl.mta.handlers.v1.ConfigurationParserTest {
 
-    private final String platformLocation;
-    private InputStream platformInputStream;
-
-    private final Expectation expectation;
-
-    private ConfigurationParser parser;
-    
     @Parameters
     public static Iterable<Object[]> getParameters() {
         return Arrays.asList(new Object[][] {
@@ -45,30 +29,12 @@ public class ConfigurationParserTest {
     }
 
     public ConfigurationParserTest(String platformsLocation, Expectation expectation) {
-        this.platformLocation = platformsLocation;
-        this.expectation = expectation;
-    }
-    
-    @Before
-    public void setUp() throws Exception {
-        if (platformLocation != null) {
-            platformInputStream = getClass().getResourceAsStream(platformLocation);
-        }
-        parser = createConfigurationParser();
+        super(platformsLocation, expectation);
     }
 
+    @Override
     protected ConfigurationParser createConfigurationParser() {
         return new ConfigurationParser();
-    }
-    
-    @Test
-    public void testParsePlatformsJson() throws Exception {
-        TestUtil.test(new Callable<Platform>() {
-            @Override
-            public Platform call() throws Exception {
-                return parser.parsePlatformJson(platformInputStream);
-            }
-        }, expectation, getClass());
     }
 
 }
