+++ com.sap.cloud.lm.sl.mta/src/test/java/com/sap/cloud/lm/sl/mta/mergers/v2/PlatformMegerTest.java
@@ -1,27 +1,19 @@
 package com.sap.cloud.lm.sl.mta.mergers.v2;
 
-import java.io.InputStream;
 import java.util.Arrays;
 import java.util.Collection;
 
-import org.junit.Before;
-import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 
-import com.sap.cloud.lm.sl.common.util.Callable;
-import com.sap.cloud.lm.sl.common.util.TestUtil;
 import com.sap.cloud.lm.sl.common.util.TestUtil.Expectation;
-import com.sap.cloud.lm.sl.mta.handlers.HandlerFactory;
-import com.sap.cloud.lm.sl.mta.handlers.v2.ConfigurationParser;
-import com.sap.cloud.lm.sl.mta.handlers.v2.DescriptorHandler;
-import com.sap.cloud.lm.sl.mta.handlers.v2.DescriptorParser;
-import com.sap.cloud.lm.sl.mta.model.v2.DeploymentDescriptor;
-import com.sap.cloud.lm.sl.mta.model.v2.Platform;
+import com.sap.cloud.lm.sl.mta.handlers.v1.DescriptorHandler;
+import com.sap.cloud.lm.sl.mta.mergers.v1.PlatformMerger;
+import com.sap.cloud.lm.sl.mta.model.v1.Platform;
 
 @RunWith(Parameterized.class)
-public class PlatformMegerTest {
+public class PlatformMegerTest extends com.sap.cloud.lm.sl.mta.mergers.v1.PlatformMergerTest {
 
     @Parameters
     public static Collection<Object[]> getParameters() {
@@ -43,61 +35,18 @@ public class PlatformMegerTest {
         });
     }
 
-    private String deploymentDescriptorLocation;
-    private String platformLocation;
-    private Expectation expectation;
-    private DeploymentDescriptor descriptor;
-    private Platform platform;
-
     public PlatformMegerTest(String deploymentDescriptorLocation, String platformTypesLocation, Expectation expectation) {
-        this.deploymentDescriptorLocation = deploymentDescriptorLocation;
-        this.platformLocation = platformTypesLocation;
-        this.expectation = expectation;
+        super(deploymentDescriptorLocation, platformTypesLocation, expectation);
     }
 
+    @Override
     protected int getMajorVersion() {
         return 2;
     }
 
+    @Override
     protected PlatformMerger getPlatformMerger(Platform platform, DescriptorHandler handler) {
-        return new PlatformMerger(platform, handler);
+        return new com.sap.cloud.lm.sl.mta.mergers.v2.PlatformMerger((com.sap.cloud.lm.sl.mta.model.v2.Platform) platform,
+            (com.sap.cloud.lm.sl.mta.handlers.v2.DescriptorHandler) handler);
     }
-
-    @Before
-    public void prepare() throws Exception {
-        loadDeploymentDescriptor();
-        loadPlatform();
-    }
-
-    private void loadPlatform() throws Exception {
-        DescriptorParser parser = getHandlerFactory().getDescriptorParser();
-        InputStream deploymentDescriptorYaml = getClass().getResourceAsStream(deploymentDescriptorLocation);
-        this.descriptor = parser.parseDeploymentDescriptorYaml(deploymentDescriptorYaml);
-    }
-
-    private void loadDeploymentDescriptor() throws Exception {
-        ConfigurationParser parser = getHandlerFactory().getConfigurationParser();
-        InputStream platformJson = getClass().getResourceAsStream(platformLocation);
-        this.platform = parser.parsePlatformJson(platformJson);
-    }
-
-    @Test
-    public void testMerge() {
-        DescriptorHandler handler = getHandlerFactory().getDescriptorHandler();
-        PlatformMerger merger = getPlatformMerger(platform, handler);
-        TestUtil.test(new Callable<DeploymentDescriptor>() {
-
-            @Override
-            public DeploymentDescriptor call() throws Exception {
-                merger.mergeInto(descriptor);
-                return descriptor;
-            }
-
-        }, expectation, getClass());
-    }
-
-    protected HandlerFactory getHandlerFactory() {
-        return new HandlerFactory(getMajorVersion());
-    }
-
 }
